Make contour creator its own worker
